feat: settings updates
