+++ src/containers/SideBar/components/Settings/Settings.styles.tsx
@@ -37,3 +37,12 @@ export const SeedWordsContainer = styled.div`
     grid-template-rows: repeat(12, 1fr);
     gap: 10px;
 `;
+
+export const HeadingContainer = styled.div`
+    display: flex;
+    justify-content: space-between;
+    align-items: center;
+    padding-bottom: 20px;
+    width: 100%;
+    border-bottom: 1px solid ${({ theme }) => theme.palette.divider};
+`;
